LayoutTests/inspector/debugger/setPauseOnExceptions-uncaught-expected.txt

CONSOLE MESSAGE: caught inline: {"line":1,"column":36}
CONSOLE MESSAGE: catchNested caught exception: {"line":3,"column":11,"sourceURL":"exception.js"}
CONSOLE MESSAGE: catchNested caught exception: {"line":8,"column":30,"sourceURL":"exception.js"}
CONSOLE MESSAGE: catchNested caught exception: "exception in host function"
CONSOLE MESSAGE: catchNested caught exception: "exception string"
CONSOLE MESSAGE: catchNested caught exception: {"line":1,"column":61}
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: NotFoundError: The object can not be found here.
CONSOLE MESSAGE: exception in host function
CONSOLE MESSAGE: exception string
CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: Error: error message
CONSOLE MESSAGE: caught inline: {"line":1,"column":36}
CONSOLE MESSAGE: catchNested caught exception: {"line":3,"column":11,"sourceURL":"exception.js"}
CONSOLE MESSAGE: catchNested caught exception: {"line":8,"column":30,"sourceURL":"exception.js"}
CONSOLE MESSAGE: catchNested caught exception: "exception in host function"
CONSOLE MESSAGE: catchNested caught exception: "exception string"
CONSOLE MESSAGE: catchNested caught exception: {"line":1,"column":61}
Debugger.setPauseOnExceptions uncaught

Debugger.setPauseOnExceptions - uncaught
PASS - did not pause on non-exception statements or caught exception statements
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
PASS - paused for each uncaught exception
PASS - did not pause on non-exception statements or caught exception statements
PASS
